fix: code style

+++ hugegraph-api/src/main/java/com/baidu/hugegraph/auth/HugeGraphAuthProxy.java
@@ -1464,11 +1464,14 @@ public final class HugeGraphAuthProxy implements HugeGraph {
 
         public Id createProject(HugeProject project) {
             this.updateCreator(project);
+            verifyUserPermission(HugePermission.WRITE, project);
             return this.authManager.createProject(project);
         }
 
         @Override
         public HugeProject deleteProject(Id id) {
+            verifyUserPermission(HugePermission.DELETE,
+                                 this.authManager.getTarget(id));
             return this.authManager.deleteProject(id);
         }
